tools/requirements.txt
# Core ML and data processing
tensorflow>=2.9.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0

# Serial communication
pyserial>=3.5

# Data handling
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional but useful
jupyter>=1.0.0
tqdm>=4.62.0